Refactor(IssueTracker): Move abstractions to sub-packages
