--- channel-api/src/main/java/com/synopsys/integration/alert/channel/api/issue/IssueTrackerIssueTransitioner.java
@@ -1,35 +1,0 @@
-/*
- * channel-api
- *
- * Copyright (c) 2021 Synopsys, Inc.
- *
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.synopsys.integration.alert.channel.api.issue;
-
-import java.io.Serializable;
-import java.util.Optional;
-
-import com.synopsys.integration.alert.channel.api.issue.model.IssueTransitionModel;
-import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueTrackerIssueResponseModel;
-import com.synopsys.integration.alert.common.exception.AlertException;
-
-public interface IssueTrackerIssueTransitioner<T extends Serializable> {
-    Optional<IssueTrackerIssueResponseModel> transitionIssue(IssueTransitionModel<T> issueTransitionModel) throws AlertException;
-
-}
