+++ channel-api/src/main/java/com/synopsys/integration/alert/channel/api/issue/IssueTrackerModelExtractor.java
@@ -28,13 +28,17 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
 
-import com.synopsys.integration.alert.channel.api.issue.model.ActionableIssueSearchResult;
-import com.synopsys.integration.alert.channel.api.issue.model.ExistingIssueDetails;
+import com.synopsys.integration.alert.channel.api.issue.convert.IssueTrackerMessageFormatter;
+import com.synopsys.integration.alert.channel.api.issue.convert.IssueTrackerSimpleMessageConverter;
+import com.synopsys.integration.alert.channel.api.issue.convert.ProjectIssueModelConverter;
 import com.synopsys.integration.alert.channel.api.issue.model.IssueCommentModel;
 import com.synopsys.integration.alert.channel.api.issue.model.IssueCreationModel;
 import com.synopsys.integration.alert.channel.api.issue.model.IssueTrackerModelHolder;
 import com.synopsys.integration.alert.channel.api.issue.model.IssueTransitionModel;
 import com.synopsys.integration.alert.channel.api.issue.model.ProjectIssueModel;
+import com.synopsys.integration.alert.channel.api.issue.search.ActionableIssueSearchResult;
+import com.synopsys.integration.alert.channel.api.issue.search.ExistingIssueDetails;
+import com.synopsys.integration.alert.channel.api.issue.search.IssueTrackerSearcher;
 import com.synopsys.integration.alert.common.enumeration.ItemOperation;
 import com.synopsys.integration.alert.common.exception.AlertException;
 import com.synopsys.integration.alert.processor.api.extract.model.ProviderMessageHolder;
