+++ channel-api/src/main/java/com/synopsys/integration/alert/channel/api/issue/IssueTrackerProcessor.java
@@ -26,6 +26,7 @@ import java.io.Serializable;
 import java.util.List;
 
 import com.synopsys.integration.alert.channel.api.issue.model.IssueTrackerModelHolder;
+import com.synopsys.integration.alert.channel.api.issue.send.IssueTrackerMessageSender;
 import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueTrackerResponse;
 import com.synopsys.integration.alert.common.exception.AlertException;
 import com.synopsys.integration.alert.processor.api.extract.model.ProviderMessageHolder;
