+++ channel-api/src/main/java/com/synopsys/integration/alert/channel/api/issue/model/IssueTransitionModel.java
@@ -25,6 +25,7 @@ package com.synopsys.integration.alert.channel.api.issue.model;
 import java.io.Serializable;
 import java.util.List;
 
+import com.synopsys.integration.alert.channel.api.issue.search.ExistingIssueDetails;
 import com.synopsys.integration.alert.common.channel.issuetracker.enumeration.IssueOperation;
 import com.synopsys.integration.alert.common.rest.model.AlertSerializableModel;
