+++ channel-api/src/main/java/com/synopsys/integration/alert/channel/api/issue/search/ExistingIssueDetails.java
@@ -20,33 +20,39 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.api.issue.model;
+package com.synopsys.integration.alert.channel.api.issue.search;
 
 import java.io.Serializable;
 
 import com.synopsys.integration.alert.common.rest.model.AlertSerializableModel;
 
-public class ProjectIssueSearchResult<T extends Serializable> extends AlertSerializableModel {
+public class ExistingIssueDetails<T extends Serializable> extends AlertSerializableModel {
     private final T issueId;
-    private final ExistingIssueDetails<T> existingIssueDetails;
-    private final ProjectIssueModel projectIssueModel;
+    private final String issueKey;
+    private final String issueSummary;
+    private final String issueLink;
 
-    public ProjectIssueSearchResult(T issueId, ExistingIssueDetails<T> existingIssueDetails, ProjectIssueModel projectIssueModel) {
+    public ExistingIssueDetails(T issueId, String issueKey, String issueSummary, String issueLink) {
         this.issueId = issueId;
-        this.existingIssueDetails = existingIssueDetails;
-        this.projectIssueModel = projectIssueModel;
+        this.issueKey = issueKey;
+        this.issueSummary = issueSummary;
+        this.issueLink = issueLink;
     }
 
     public T getIssueId() {
         return issueId;
     }
 
-    public ExistingIssueDetails<T> getExistingIssueDetails() {
-        return existingIssueDetails;
+    public String getIssueKey() {
+        return issueKey;
     }
 
-    public ProjectIssueModel getProjectIssueModel() {
-        return projectIssueModel;
+    public String getIssueSummary() {
+        return issueSummary;
+    }
+
+    public String getIssueLink() {
+        return issueLink;
     }
 
 }
