+++ channel-api/src/main/java/com/synopsys/integration/alert/channel/api/issue/search/IssueTrackerSearcher.java
@@ -20,16 +20,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.api.issue;
+package com.synopsys.integration.alert.channel.api.issue.search;
 
 import java.io.Serializable;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.synopsys.integration.alert.channel.api.issue.model.ActionableIssueSearchResult;
+import com.synopsys.integration.alert.channel.api.issue.convert.ProjectMessageToIssueModelTransformer;
 import com.synopsys.integration.alert.channel.api.issue.model.ProjectIssueModel;
-import com.synopsys.integration.alert.channel.api.issue.model.ProjectIssueSearchResult;
 import com.synopsys.integration.alert.common.enumeration.ItemOperation;
 import com.synopsys.integration.alert.common.exception.AlertException;
 import com.synopsys.integration.alert.common.exception.AlertRuntimeException;
