+++ channel-api/src/main/java/com/synopsys/integration/alert/channel/api/issue/send/IssueTrackerIssueCommenter.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.api.issue;
+package com.synopsys.integration.alert.channel.api.issue.send;
 
 import java.io.Serializable;
 import java.util.Optional;
