+++ channel-api/src/main/java/com/synopsys/integration/alert/channel/api/issue/send/IssueTrackerIssueTransitioner.java
@@ -20,39 +20,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.api.issue.model;
+package com.synopsys.integration.alert.channel.api.issue.send;
 
 import java.io.Serializable;
+import java.util.Optional;
 
-import com.synopsys.integration.alert.common.rest.model.AlertSerializableModel;
+import com.synopsys.integration.alert.channel.api.issue.model.IssueTransitionModel;
+import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueTrackerIssueResponseModel;
+import com.synopsys.integration.alert.common.exception.AlertException;
 
-public class ExistingIssueDetails<T extends Serializable> extends AlertSerializableModel {
-    private final T issueId;
-    private final String issueKey;
-    private final String issueSummary;
-    private final String issueLink;
-
-    public ExistingIssueDetails(T issueId, String issueKey, String issueSummary, String issueLink) {
-        this.issueId = issueId;
-        this.issueKey = issueKey;
-        this.issueSummary = issueSummary;
-        this.issueLink = issueLink;
-    }
-
-    public T getIssueId() {
-        return issueId;
-    }
-
-    public String getIssueKey() {
-        return issueKey;
-    }
-
-    public String getIssueSummary() {
-        return issueSummary;
-    }
-
-    public String getIssueLink() {
-        return issueLink;
-    }
+public interface IssueTrackerIssueTransitioner<T extends Serializable> {
+    Optional<IssueTrackerIssueResponseModel> transitionIssue(IssueTransitionModel<T> issueTransitionModel) throws AlertException;
 
 }
